페이지들 pages폴더로 이동 했고 승현이꺼 넣어볼라 했더니 스타일이 아무것도 없네?
