+++ src/pages/CommunityWrite.js
@@ -1,0 +1,31 @@
+import * as styleD from '../styles/Communuty';
+import Navbar from '../components/Navbar';
+import Line from '../components/Line';
+import { Link } from "react-router-dom";
+
+export default function MypageStorage() {
+  return (
+    <div>
+      <Navbar />
+      <Line /> 
+      <styleD.CommunityHeaderContainer>
+        <styleD.CommunitySelectBox>
+            <option>제목</option>
+            <option>내용</option>
+            <option>작성자</option>
+        </styleD.CommunitySelectBox>
+
+          <styleD.CommunitySearch className='search-input' placeholder='키워드를 입력해 주세요' />
+          <styleD.HeaderSearchBtn><img className='search-btn' src='./source/Icon.png' /></styleD.HeaderSearchBtn>
+      </styleD.CommunityHeaderContainer>
+
+      <Line />
+    
+    
+    
+
+
+    </div>
+  );
+}
+
