+++ src/pages/MypageStorage.js
@@ -1,6 +1,6 @@
 import * as styleD from '../styles/Mypage';
-import Navbar from './Navbar';
-import Line from './Line';
+import Navbar from '../components/Navbar';
+import Line from '../components/Line';
 import { Link } from "react-router-dom";
 
 export default function MypageStorage() {
